feat: 事件触发参数通用化 #5620
