+++ src/backend/ci/core/common/common-webhook/api-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/pojo/code/WebhookConstant.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.scm.pojo
+package com.tencent.devops.common.webhook.pojo.code
 
 const val BK_CI_RUN = "BK_CI_RUN"
 const val BK_CI_REF = "BK_CI_REF"
@@ -140,3 +140,17 @@ const val BK_REPO_P4_WEBHOOK_EVENT_TYPE = "BK_REPO_P4_WEBHOOK_EVENT_TYPE"
 const val BK_REPO_P4_WEBHOOK_INCLUDE_PATHS = "BK_REPO_P4_WEBHOOK_INCLUDE_PATHS"
 const val BK_REPO_P4_WEBHOOK_EXCLUDE_PATHS = "BK_REPO_P4_WEBHOOK_EXCLUDE_PATHS"
 const val BK_REPO_P4_WEBHOOK_CHANGE = "BK_REPO_P4_WEBHOOK_CHANGE"
+
+const val PIPELINE_WEBHOOK_REVISION = "BK_CI_HOOK_REVISION" // hookRevision
+const val PIPELINE_WEBHOOK_BRANCH = "BK_CI_HOOK_BRANCH" // hookBranch
+const val PIPELINE_WEBHOOK_SOURCE_URL = "BK_CI_HOOK_SOURCE_URL" // hookSourceUrl
+const val PIPELINE_WEBHOOK_TARGET_URL = "BK_CI_HOOK_TARGET_URL" // hookTargetUrl
+const val PIPELINE_WEBHOOK_REPO = "BK_CI_HOOK_REPO" // hookRepo
+const val PIPELINE_WEBHOOK_REPO_TYPE = "BK_CI_HOOK_REPO_TYPE" // hookRepoType
+const val PIPELINE_WEBHOOK_BLOCK = "BK_CI_HOOK_BLOCK" // hookBlock
+const val PIPELINE_WEBHOOK_TYPE = "BK_CI_HOOK_TYPE" // hookType
+const val PIPELINE_WEBHOOK_EVENT_TYPE = "BK_CI_HOOK_EVENT_TYPE" // hookEventType
+const val PIPELINE_REPO_NAME = "BK_CI_REPO_NAME" // "repoName"
+const val PIPELINE_WEBHOOK_COMMIT_MESSAGE = "BK_CI_HOOK_MESSAGE" // hook message
+const val PIPELINE_START_WEBHOOK_USER_ID = "BK_CI_START_WEBHOOK_USER_ID" // "pipeline.start.webhook.user.id"
+const val PIPELINE_WEBHOOK_QUEUE = "BK_CI_HOOK_QUEUE"
