+++ src/backend/ci/core/common/common-webhook/biz-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/service/code/filter/BranchFilter.kt
@@ -27,7 +27,7 @@
 
 package com.tencent.devops.common.webhook.service.code.filter
 
-import com.tencent.devops.scm.pojo.MATCH_BRANCH
+import com.tencent.devops.common.webhook.pojo.code.MATCH_BRANCH
 import org.slf4j.LoggerFactory
 import org.springframework.util.AntPathMatcher
