+++ src/backend/ci/core/common/common-webhook/biz-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/service/code/param/GitlabWebhookElementParams.kt
@@ -25,40 +25,60 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.process.engine.service.code
+package com.tencent.devops.common.webhook.service.code.param
 
 import com.tencent.devops.common.api.util.EnvUtils
-import com.tencent.devops.common.pipeline.pojo.element.trigger.CodeSVNWebHookTriggerElement
+import com.tencent.devops.common.pipeline.pojo.element.trigger.CodeGitlabWebHookTriggerElement
 import com.tencent.devops.common.pipeline.pojo.element.trigger.enums.CodeType
 import com.tencent.devops.common.pipeline.utils.RepositoryConfigUtils
 import com.tencent.devops.common.webhook.pojo.code.WebHookParams
-import com.tencent.devops.process.pojo.code.ScmWebhookElementParams
+import org.springframework.stereotype.Service
 
-class SvnWebhookElementParams : ScmWebhookElementParams<CodeSVNWebHookTriggerElement> {
+@Service
+class GitlabWebhookElementParams : ScmWebhookElementParams<CodeGitlabWebHookTriggerElement> {
+
+    override fun elementClass(): Class<CodeGitlabWebHookTriggerElement> {
+        return CodeGitlabWebHookTriggerElement::class.java
+    }
+
+    @SuppressWarnings("ComplexMethod")
     override fun getWebhookElementParams(
-        element: CodeSVNWebHookTriggerElement,
+        element: CodeGitlabWebHookTriggerElement,
         variables: Map<String, String>
-    ): WebHookParams {
+    ): WebHookParams? {
         val params = WebHookParams(
             repositoryConfig = RepositoryConfigUtils.replaceCodeProp(
                 repositoryConfig = RepositoryConfigUtils.buildConfig(element),
                 variables = variables
             )
         )
-        params.pathFilterType = element.pathFilterType
-        params.relativePath = EnvUtils.parseEnv(element.relativePath ?: "", variables)
-        params.excludeUsers = if (element.excludeUsers == null || element.excludeUsers!!.isEmpty()) {
-            ""
-        } else {
-            EnvUtils.parseEnv(element.excludeUsers!!.joinToString(","), variables)
-        }
         params.includeUsers = if (element.includeUsers == null || element.includeUsers!!.isEmpty()) {
             ""
         } else {
             EnvUtils.parseEnv(element.includeUsers!!.joinToString(","), variables)
         }
+        params.excludeUsers = if (element.excludeUsers == null || element.excludeUsers!!.isEmpty()) {
+            ""
+        } else {
+            EnvUtils.parseEnv(element.excludeUsers!!.joinToString(","), variables)
+        }
+        if (element.branchName == null) {
+            return null
+        }
+        params.branchName = EnvUtils.parseEnv(element.branchName!!, variables)
+        params.codeType = CodeType.GITLAB
+        params.eventType = element.eventType
+        params.block = element.block ?: false
+        params.excludeBranchName = EnvUtils.parseEnv(element.excludeBranchName ?: "", variables)
+        params.pathFilterType = element.pathFilterType
+        params.includePaths = EnvUtils.parseEnv(element.includePaths ?: "", variables)
         params.excludePaths = EnvUtils.parseEnv(element.excludePaths ?: "", variables)
-        params.codeType = CodeType.SVN
+        params.tagName = EnvUtils.parseEnv(element.tagName ?: "", variables)
+        params.excludeTagName = EnvUtils.parseEnv(element.excludeTagName ?: "", variables)
+        params.excludeSourceBranchName = EnvUtils.parseEnv(element.excludeSourceBranchName ?: "", variables)
+        params.includeSourceBranchName = EnvUtils.parseEnv(element.includeSourceBranchName ?: "", variables)
+        params.includeCommitMsg = EnvUtils.parseEnv(element.includeCommitMsg ?: "", variables)
+        params.excludeCommitMsg = EnvUtils.parseEnv(element.excludeCommitMsg ?: "", variables)
         return params
     }
 }
