+++ src/backend/ci/core/common/common-webhook/biz-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/service/code/param/SvnWebHookStartParam.kt
@@ -25,40 +25,41 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.process.engine.service.code
+package com.tencent.devops.common.webhook.service.code.param
 
-import com.tencent.devops.common.pipeline.pojo.element.trigger.CodeP4WebHookTriggerElement
+import com.tencent.devops.common.pipeline.pojo.element.trigger.CodeSVNWebHookTriggerElement
 import com.tencent.devops.common.webhook.pojo.code.WebHookParams
-import com.tencent.devops.common.webhook.service.code.matcher.P4WebHookMatcher
 import com.tencent.devops.common.webhook.service.code.matcher.ScmWebhookMatcher
-import com.tencent.devops.process.pojo.code.ScmWebhookStartParams
+import com.tencent.devops.common.webhook.service.code.matcher.SvnWebHookMatcher
 import com.tencent.devops.repository.pojo.Repository
-import com.tencent.devops.scm.pojo.BK_REPO_P4_WEBHOOK_EVENT_TYPE
-import com.tencent.devops.scm.pojo.BK_REPO_P4_WEBHOOK_INCLUDE_PATHS
-import com.tencent.devops.scm.pojo.BK_REPO_P4_WEBHOOK_P4PORT
+import com.tencent.devops.common.webhook.pojo.code.BK_REPO_SVN_WEBHOOK_EXCLUDE_PATHS
+import com.tencent.devops.common.webhook.pojo.code.BK_REPO_SVN_WEBHOOK_EXCLUDE_USERS
+import com.tencent.devops.common.webhook.pojo.code.BK_REPO_SVN_WEBHOOK_INCLUDE_USERS
+import com.tencent.devops.common.webhook.pojo.code.BK_REPO_SVN_WEBHOOK_RELATIVE_PATH
 
-class P4WebHookStartParam(
-    val projectId: String,
-    val repo: Repository,
-    private val params: WebHookParams,
-    private val matcher: P4WebHookMatcher,
-    private val matchResult: ScmWebhookMatcher.MatchResult
-) : ScmWebhookStartParams<CodeP4WebHookTriggerElement> {
+class SvnWebHookStartParam(
+    private val matcher: SvnWebHookMatcher
+) : ScmWebhookStartParams<CodeSVNWebHookTriggerElement> {
 
-    override fun getStartParams(element: CodeP4WebHookTriggerElement): Map<String, Any> {
+    override fun elementClass(): Class<CodeSVNWebHookTriggerElement> {
+        return CodeSVNWebHookTriggerElement::class.java
+    }
+
+    override fun getElementStartParams(
+        projectId: String,
+        element: CodeSVNWebHookTriggerElement,
+        repo: Repository,
+        matcher: ScmWebhookMatcher,
+        variables: Map<String, String>,
+        params: WebHookParams,
+        matchResult: ScmWebhookMatcher.MatchResult
+    ): Map<String, Any> {
         val startParams = mutableMapOf<String, Any>()
-        with(element.data.input) {
-            startParams[BK_REPO_P4_WEBHOOK_P4PORT] = repo.url
-            startParams[BK_REPO_P4_WEBHOOK_EVENT_TYPE] = params.eventType ?: ""
-            startParams[BK_REPO_P4_WEBHOOK_INCLUDE_PATHS] = includePaths ?: ""
-            startParams.putAll(
-                matcher.retrieveParams(
-                    projectId = projectId,
-                    repository = repo
-                )
-            )
-        }
-
+        startParams[BK_REPO_SVN_WEBHOOK_RELATIVE_PATH] = element.relativePath ?: ""
+        startParams[BK_REPO_SVN_WEBHOOK_EXCLUDE_PATHS] = element.excludePaths ?: ""
+        startParams[BK_REPO_SVN_WEBHOOK_INCLUDE_USERS] = element.includeUsers?.joinToString(",") ?: ""
+        startParams[BK_REPO_SVN_WEBHOOK_EXCLUDE_USERS] = element.excludeUsers?.joinToString(",") ?: ""
+        startParams.putAll(matcher.retrieveParams())
         return startParams
     }
 }
