+++ src/backend/ci/core/process/biz-base/src/test/kotlin/com/tencent/devops/process/engine/compatibility/v2/V2BuildParametersCompatibilityTransformerTest.kt
@@ -32,11 +32,11 @@ import com.nhaarman.mockito_kotlin.whenever
 import com.tencent.devops.common.pipeline.enums.BuildFormPropertyType
 import com.tencent.devops.common.pipeline.pojo.BuildFormProperty
 import com.tencent.devops.common.pipeline.pojo.BuildParameters
+import com.tencent.devops.common.webhook.pojo.code.PIPELINE_WEBHOOK_BRANCH
 import com.tencent.devops.process.util.PswParameterUtils
 import com.tencent.devops.process.utils.FIXVERSION
 import com.tencent.devops.process.utils.MAJORVERSION
 import com.tencent.devops.process.utils.MINORVERSION
-import com.tencent.devops.process.utils.PIPELINE_WEBHOOK_BRANCH
 import com.tencent.devops.process.utils.PipelineVarUtil
 import org.junit.Assert
 import org.junit.Assert.assertEquals
